@@ -11,20 +11,20 @@ members = [
 ]
 
 [workspace.package]
-version = "0.35.0"
+version = "0.38.2"
 authors = ["DFINITY Stiftung <sdk@dfinity.org>"]
 edition = "2021"
 repository = "https://github.com/dfinity/agent-rs"
 # MSRV
 # Avoid updating this field unless we use new Rust features
 # Sync rust-version in rust-toolchain.toml
-rust-version = "1.70.0"
+rust-version = "1.75.0"
 license = "Apache-2.0"
 
 [workspace.dependencies]
-ic-agent = { path = "ic-agent", version = "0.35.0", default-features = false }
-ic-utils = { path = "ic-utils", version = "0.35.0" }
-ic-transport-types = { path = "ic-transport-types", version = "0.35.0" }
+ic-agent = { path = "ic-agent", version = "0.38.2", default-features = false }
+ic-utils = { path = "ic-utils", version = "0.38.2" }
+ic-transport-types = { path = "ic-transport-types", version = "0.38.2" }
 
 ic-certification = "2.2"
 candid = "0.10.1"
@@ -33,6 +33,7 @@ clap = "4.4.3"
 futures-util = "0.3.21"
 hex = "0.4.3"
 leb128 = "0.2.5"
+reqwest = { version = "0.12", default-features = false }
 ring = "0.17.7"
 serde = "1.0.162"
 serde_bytes = "0.11.13"
