@@ -1,8 +1,12 @@
+use url::Url;
+
 use crate::{
-    agent::{agent_config::AgentConfig, Agent, Transport},
+    agent::{agent_config::AgentConfig, Agent},
     AgentError, Identity, NonceFactory, NonceGenerator,
 };
 use std::sync::Arc;
+
+use super::{route_provider::RouteProvider, HttpService};
 
 /// A builder for an [`Agent`].
 #[derive(Default)]
@@ -16,24 +20,34 @@ impl AgentBuilder {
         Agent::new(self.config)
     }
 
+    /// Set the dynamic transport layer for the [`Agent`], performing continuous discovery of the API boundary nodes and routing traffic via them based on latency.
+    #[cfg(feature = "_internal_dynamic-routing")]
+    pub async fn with_discovery_transport(self, client: reqwest::Client) -> Self {
+        use crate::agent::route_provider::dynamic_routing::{
+            dynamic_route_provider::{DynamicRouteProviderBuilder, IC0_SEED_DOMAIN},
+            node::Node,
+            snapshot::latency_based_routing::LatencyRoutingSnapshot,
+        };
+        // TODO: This is a temporary solution to get the seed node.
+        let seed = Node::new(IC0_SEED_DOMAIN).unwrap();
+
+        let route_provider = DynamicRouteProviderBuilder::new(
+            LatencyRoutingSnapshot::new(),
+            vec![seed],
+            client.clone(),
+        )
+        .build()
+        .await;
+
+        let route_provider = Arc::new(route_provider) as Arc<dyn RouteProvider>;
+
+        self.with_arc_route_provider(route_provider)
+            .with_http_client(client)
+    }
+
     /// Set the URL of the [Agent].
-    #[cfg(feature = "reqwest")]
     pub fn with_url<S: Into<String>>(self, url: S) -> Self {
-        use crate::agent::http_transport::ReqwestTransport;
-
-        self.with_transport(ReqwestTransport::create(url).unwrap())
-    }
-
-    /// Set a Replica transport to talk to serve as the replica interface.
-    pub fn with_transport<T: 'static + Transport>(self, transport: T) -> Self {
-        self.with_arc_transport(Arc::new(transport))
-    }
-
-    /// Same as [Self::with_transport], but provides a `Arc` boxed implementation instead
-    /// of a direct type.
-    pub fn with_arc_transport(mut self, transport: Arc<dyn Transport>) -> Self {
-        self.config.transport = Some(transport);
-        self
+        self.with_route_provider(url.into().parse::<Url>().unwrap())
     }
 
     /// Add a NonceFactory to this Agent. By default, no nonce is produced.
@@ -41,7 +55,7 @@ impl AgentBuilder {
         self.with_nonce_generator(nonce_factory)
     }
 
-    /// Same as [Self::with_nonce_factory], but for any `NonceGenerator` type
+    /// Same as [`Self::with_nonce_factory`], but for any `NonceGenerator` type
     pub fn with_nonce_generator<N: 'static + NonceGenerator>(
         self,
         nonce_factory: N,
@@ -49,8 +63,7 @@ impl AgentBuilder {
         self.with_arc_nonce_generator(Arc::new(nonce_factory))
     }
 
-    /// Same as [Self::with_nonce_generator], but provides a `Arc` boxed implementation instead
-    /// of a direct type.
+    /// Same as [`Self::with_nonce_generator`], but reuses an existing `Arc`.
     pub fn with_arc_nonce_generator(
         mut self,
         nonce_factory: Arc<dyn NonceGenerator>,
@@ -67,14 +80,12 @@ impl AgentBuilder {
         self.with_arc_identity(Arc::new(identity))
     }
 
-    /// Same as [Self::with_identity], but provides a boxed implementation instead
-    /// of a direct type.
+    /// Same as [`Self::with_identity`], but reuses an existing box
     pub fn with_boxed_identity(self, identity: Box<dyn Identity>) -> Self {
         self.with_arc_identity(Arc::from(identity))
     }
 
-    /// Same as [Self::with_identity], but provides a `Arc` boxed implementation instead
-    /// of a direct type.
+    /// Same as [`Self::with_identity`], but reuses an existing `Arc`
     pub fn with_arc_identity(mut self, identity: Arc<dyn Identity>) -> Self {
         self.config.identity = identity;
         self
@@ -104,6 +115,53 @@ impl AgentBuilder {
     /// to avoid the slowdown of retrying any 429 errors.
     pub fn with_max_concurrent_requests(mut self, max_concurrent_requests: usize) -> Self {
         self.config.max_concurrent_requests = max_concurrent_requests;
+        self
+    }
+
+    /// Add a `RouteProvider` to this agent, to provide the URLs of boundary nodes.
+    pub fn with_route_provider(self, provider: impl RouteProvider + 'static) -> Self {
+        self.with_arc_route_provider(Arc::new(provider))
+    }
+
+    /// Same as [`Self::with_route_provider`], but reuses an existing `Arc`.
+    pub fn with_arc_route_provider(mut self, provider: Arc<dyn RouteProvider>) -> Self {
+        self.config.route_provider = Some(provider);
+        self
+    }
+
+    /// Provide a pre-configured HTTP client to use. Use this to set e.g. HTTP timeouts or proxy configuration.
+    pub fn with_http_client(mut self, client: reqwest::Client) -> Self {
+        assert!(
+            self.config.http_service.is_none(),
+            "with_arc_http_middleware cannot be called with with_http_client"
+        );
+        self.config.client = Some(client);
+        self
+    }
+
+    /// Provide a custom `reqwest`-compatible HTTP service, e.g. to add per-request headers for custom boundary nodes.
+    /// Most users will not need this and should use `with_http_client`. Cannot be called with `with_http_client`.
+    ///
+    /// The trait is automatically implemented for any `tower::Service` impl matching the one `reqwest::Client` uses,
+    /// including `reqwest-middleware`. This is a low-level interface, and direct implementations must provide all automatic retry logic.
+    pub fn with_arc_http_middleware(mut self, service: Arc<dyn HttpService>) -> Self {
+        assert!(
+            self.config.client.is_none(),
+            "with_arc_http_middleware cannot be called with with_http_client"
+        );
+        self.config.http_service = Some(service);
+        self
+    }
+
+    /// Retry up to the specified number of times upon encountering underlying TCP errors.
+    pub fn with_max_tcp_error_retries(mut self, retries: usize) -> Self {
+        self.config.max_tcp_error_retries = retries;
+        self
+    }
+
+    /// Don't accept HTTP bodies any larger than `max_size` bytes.
+    pub fn with_max_response_body_size(mut self, max_size: usize) -> Self {
+        self.config.max_response_body_size = Some(max_size);
         self
     }
 }
