@@ -409,7 +409,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         >,
         body: impl AsRef<[u8]>,
         certificate_version: Option<&u16>,
-    ) -> impl 'agent + SyncCall<(HttpResponse,)> {
+    ) -> impl 'agent + SyncCall<Value = (HttpResponse,)> {
         self.http_request_custom(
             method.as_ref(),
             url.as_ref(),
@@ -428,7 +428,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         headers: H,
         body: &[u8],
         certificate_version: Option<&u16>,
-    ) -> impl 'agent + SyncCall<(HttpResponse<T, C>,)>
+    ) -> impl 'agent + SyncCall<Value = (HttpResponse<T, C>,)>
     where
         H: 'agent + Send + Sync + Clone + ExactSizeIterator<Item = HeaderField<'agent>>,
         T: 'agent + Send + Sync + CandidType + for<'de> Deserialize<'de>,
@@ -453,7 +453,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         url: impl AsRef<str>,
         headers: impl 'agent + Send + Sync + Clone + ExactSizeIterator<Item = HeaderField<'agent>>,
         body: impl AsRef<[u8]>,
-    ) -> impl 'agent + AsyncCall<(HttpResponse,)> {
+    ) -> impl 'agent + AsyncCall<Value = (HttpResponse,)> {
         self.http_request_update_custom(method.as_ref(), url.as_ref(), headers, body.as_ref())
     }
 
@@ -466,7 +466,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         url: &str,
         headers: H,
         body: &[u8],
-    ) -> impl 'agent + AsyncCall<(HttpResponse<T, C>,)>
+    ) -> impl 'agent + AsyncCall<Value = (HttpResponse<T, C>,)>
     where
         H: 'agent + Send + Sync + Clone + ExactSizeIterator<Item = HeaderField<'agent>>,
         T: 'agent + Send + Sync + CandidType + for<'de> Deserialize<'de>,
@@ -487,7 +487,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         &'canister self,
         method: impl AsRef<str>,
         token: Token,
-    ) -> impl 'agent + SyncCall<(StreamingCallbackHttpResponse,)> {
+    ) -> impl 'agent + SyncCall<Value = (StreamingCallbackHttpResponse,)> {
         self.query(method.as_ref()).with_value_arg(token.0).build()
     }
 
@@ -497,7 +497,7 @@ impl<'agent> HttpRequestCanister<'agent> {
         &'canister self,
         method: impl AsRef<str>,
         token: T,
-    ) -> impl 'agent + SyncCall<(StreamingCallbackHttpResponse<T>,)>
+    ) -> impl 'agent + SyncCall<Value = (StreamingCallbackHttpResponse<T>,)>
     where
         T: 'agent + Send + Sync + CandidType + for<'de> Deserialize<'de>,
     {
